move runCommand to testutil package
